refactor(ios): extract the upload route from MediaUploadServer
The upload handler and everything it calls move into MediaUploadRoute,
a LocalServerRoute that owns the upload delegate and default uploader.
The route holds the delegate weakly, as the upload context did, for the
same retain-cycle reason.

The server becomes a shell: it builds the route list — relay first, so
its prefix is checked ahead of the upload route's exact path — and its
handler hands each request to the first route that claims it. The
constructor, the server-level refusals, and every test are unchanged.
